In order to support tags with parameter
I really recommend supporting tags with parameters:

I would like to either have a tag or an attribute as element.

The control flow of the application needs to distinct between tags or attributes anyway.
I could either have a single tag element with attributes or an attribute element with further attributes. 

This change is just a hot fix for recognizing such tags as well and let the client application handle that case which works fine for the moment but should be replaced by a clean structure.